agenda_pyqt/backend/requirements.txt
fastapi>=0.109.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
uvicorn>=0.25.0
sqlalchemy>=2.0.25
alembic>=1.13.1
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.1,<3.0.0
